requirements.txt
aiogram
telethon
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
httpx
feedparser
rapidfuzz
apscheduler
python-dotenv